+++ docs/conf.py
@@ -47,6 +47,7 @@ ipython_execlines = [
     "pd.set_option('display.width', 1000000)",
     "pd.set_option('max_columns', 1000)",
 ]
+autoclass_content = 'both'
 autosummary_generate = True
 autodoc_typehints = "none"
 
@@ -134,8 +135,7 @@ html_context = {
 #     'display_version': False,
 # }
 html_theme_options = {
-    "github_url": "https://github.com/sdv-dev/fibber",
-    # "twitter_url": "https://twitter.com/sdv_dev",
+    "github_url": "https://github.com/dai-lab/fibber",
     "show_prev_next": True,
 }