+++ fibber/metrics/metric_base.py
@@ -3,16 +3,16 @@ class MetricBase(object):
 
     All metrics should be derived from this class.
 
-    To implement a new metric, you should at least overwrite the `__call__` method.
+    To implement a new metric, you should at least overwrite the `measure_example` method.
 
     The simplest metric can be directly computed from a pair of text, in this case, the metric
     can use the `origin` and `paraphrase` args directly.
 
     Other metrics need more information from the data record. For example, `text0`, `text1`, or
-    `label`. Thus the `data_record` and `paraphrase_field` are also provided as args.
+    ``label``. Thus the `data_record` and `paraphrase_field` are also provided as args.
 
     Some metrics may run more efficiently on a batch of data. In this case, you should overwrite
-    the batch_call function. If you don't overwrite batch_call, it will compute the metric of
+    the ``measure_batch`` function. If you don't overwrite batch_call, it will compute the metric of
     paraphrases one by one.
     """
 
@@ -22,7 +22,7 @@ class MetricBase(object):
     def __repr__(self):
         return self.__class__.__name__
 
-    def batch_call(self, origin, paraphrases, data_record=None, paraphrase_field="text0"):
+    def measure_batch(self, origin, paraphrases, data_record=None, paraphrase_field="text0"):
         """Measure the metric on a batch of paraphrases.
 
         Args:
@@ -36,8 +36,8 @@ class MetricBase(object):
         """
         ret = []
         for paraphrase in paraphrases:
-            ret.append(self(origin, paraphrase, data_record, paraphrase_field))
+            ret.append(self.measure_example(origin, paraphrase, data_record, paraphrase_field))
         return ret
 
-    def __call__(self, origin, paraphrase, data_record=None, paraphrase_field="text0"):
+    def measure_example(self, origin, paraphrase, data_record=None, paraphrase_field="text0"):
         raise NotImplementedError()