+++ fibber/metrics/metric_utils.py
@@ -114,7 +114,7 @@ class MetricBundle(object):
         """Returns the classifier for attack."""
         return self.get_metric("BertClfPrediction")
 
-    def __call__(self, origin, paraphrase, data_record=None, paraphrase_field="text0"):
+    def measure_example(self, origin, paraphrase, data_record=None, paraphrase_field="text0"):
         """Compute the results of all metrics in the bundle for one pair of text.
 
         Args:
@@ -128,7 +128,7 @@ class MetricBundle(object):
         """
         ret = {}
         for name, metric in self._metrics.items():
-            ret[name] = metric(origin, paraphrase, data_record, paraphrase_field)
+            ret[name] = metric.measure_example(origin, paraphrase, data_record, paraphrase_field)
         return ret
 
 
@@ -151,7 +151,7 @@ def compute_metrics(metric_bundle, results, output_filename):
         data_record_tmp = dict([(k, v) for k, v in data_record.items() if "_paraphrases" not in k])
 
         # Run metrics on original text
-        data_record["original_text_metrics"] = metric_bundle(
+        data_record["original_text_metrics"] = metric_bundle.measure_example(
             data_record[paraphrase_field], data_record[paraphrase_field],
             data_record_tmp, paraphrase_field)
 
@@ -159,8 +159,8 @@ def compute_metrics(metric_bundle, results, output_filename):
         paraphrase_metric_list = []
         for paraphrase in data_record[paraphrase_field + "_paraphrases"]:
             paraphrase_metric_list.append(
-                metric_bundle(data_record[paraphrase_field], paraphrase, data_record_tmp,
-                              paraphrase_field))
+                metric_bundle.measure_example(
+                    data_record[paraphrase_field], paraphrase, data_record_tmp, paraphrase_field))
 
         data_record["paraphrase_metrics"] = paraphrase_metric_list
 
@@ -178,7 +178,7 @@ def compute_metrics(metric_bundle, results, output_filename):
 
 def aggregate_metrics(dataset_name, paraphrase_strategy_name, experiment_name, results,
                       customize_aggregation_fn_dict):
-    """Aggregate paraphrase metrics on a dataset to a summary and store in a dictionary.
+    """Aggregate paraphrase metrics on a dataset to a summary and store in a dict.
 
     Args:
         dataset_name (str): the name of the dataset.
@@ -186,7 +186,7 @@ def aggregate_metrics(dataset_name, paraphrase_strategy_name, experiment_name, r
         experiment_name (str): the name of the experiment.
         results (dict): the fibber dataset with paraphrases and metrics. The return value of
             `compute_metrics`.
-        customize_aggregation_fn_dict (dict): A dict of customized aggregations. The dictionary is
+        customize_aggregation_fn_dict (dict): A dict of customized aggregations. The dict is
             a mapping from aggregation name to aggregation function. The aggregation function
             should take one data_record, and returns a float.