+++ fibber/resources/__init__.py
@@ -1,0 +1,3 @@
+from fibber.resources.resource_utils import get_glove_emb, get_stopwords
+
+__all__ = ["get_glove_emb", "get_stopwords"]