+++ HISTORY.md
@@ -1,1 +1,15 @@
 # History
+
+## version 0.0.1
+
+This is the first release of Fibber library. This release contains:
+
+- Datasets: fibber contains 6 built-in datasets.
+- Metrics: fibber contains 6 metrics to evaluate the quality of paraphrased
+  sentences. All metrics have a unified interface.
+- Benchmark framework: the benchmark framework and easily evaluate the
+  phraphrase strategies on built-in datasets and metrics.
+- Strategies: this release contains 2 basic strategies, the identity strategy
+  and random strategy.
+- A unified Fibber interface: users can easily use fibber by creating a Fibber
+  object.